// log-manager-front/src/lib/api/query/mutationKey.ts
export enum MUTATION_KEY {
  ACCOUNT_ADD_USER = 'ACCOUNT_ADD_USER',
  ACCOUNT_CHANGE_PASSWORD = 'ACCOUNT_ADD_USER',
  ACCOUNT_EDIT_PERMISSION = 'ACCOUNT_EDIT_PERMISSION',
  AUTH_LOGIN = 'AUTH_LOGIN',
  AUTH_LOGOUT = 'AUTH_LOGOUT',
  AUTH_REISSUE_TOKEN = 'AUTH_REISSUE_TOKEN',
  CONFIGURE_TEST_CRAS_SERVER = 'CONFIGURE_TEST_CRAS_SERVER',
  CONFIGURE_TEST_EMAIL_SERVER = 'CONFIGURE_TEST_EMAIL_SERVER',
  CONFIGURE_TEST_RAPID_COLLECTOR_SERVER = 'CONFIGURE_TEST_RAPID_COLLECTOR_SERVER',
  CONFIGURE_ADD_SITE = 'CONFIGURE_ADD_SITE',
  CONFIGURE_EDIT_SITE = 'CONFIGURE_EDIT_SITE',
  ADDRESS_ADD_EDIT_EMAIL = 'ADDRESS_ADD_EDIT_EMAIL',
  ADDRESS_ADD_EDIT_GROUP = 'ADDRESS_ADD_EDIT_GROUP',
  ADDRESS_DELETE_GROUP = 'ADDRESS_DELETE_GROUP',
  ADDRESS_SEARCH_EMAIL = 'ADDRESS_SEARCH_EMAIL',
  ADDRESS_SEARCH_EMAIL_GROUP = 'ADDRESS_SEARCH_EMAIL_GROUP',
  ADDRESS_DELETE_EMAIL = 'ADDRESS_DELETE_EMAIL',

  RULES_CRAS_ADD_SITE = 'RULES_CRAS_ADD_SITE',
  RULES_CRAS_DELETE = 'RULES_CRAS_DELETE',
  RULES_CRAS_CREATE_TEST_QUERY = 'RULES_CRAS_CREATE_TEST_QUERY',
  RULES_CRAS_CREATE_ADD = 'RULES_CRAS_CREATE_ADD',
  RULES_CRAS_CREATE_EDIT = 'RULES_CRAS_CREATE_EDIT',
  RULES_CRAS_CREATE_DELETE = 'RULES_CRAS_CREATE_DELETE',
  RULES_CRAS_JUDGE_ADD = 'RULES_CRAS_JUDGE_ADD',
  RULES_CRAS_JUDGE_EDIT = 'RULES_CRAS_JUDGE_EDIT',
  RULES_CRAS_JUDGE_DELETE = 'RULES_CRAS_JUDGE_DELETE',
}